% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_one}
\alias{add_one}
\title{Add One to a Number}
\usage{
add_one(number)
}
\arguments{
\item{number}{Numeric. A number to which one will be added. See details.}
}
\value{
Numeric.
}
\description{
Add One to a Number
}
\details{
At MEMBER LEVEL 1, you'll receive the answer, but it'll be wrong 20\% of the
time!

At MEMBER LEVEL 2, the answer will always be correct!

At MEMBER LEVEL 3, the answer will print correctly... with a hat!

If you run out of MEGA COINS, simply \link{watch_ad}!
}
\examples{
\dontrun{add_one(1)}
}
