% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads.R
\name{watch_ad}
\alias{watch_ad}
\title{Watch an Ad to Receive More MEGA COINS}
\usage{
watch_ad(duration = c("short", "long"))
}
\arguments{
\item{duration}{Character. Length of the ad, either 'short' or 'long'.}
}
\value{
Nothing.
}
\description{
Watch an Ad to Receive More MEGA COINS
}
\examples{
\dontrun{watch_ad("long")}
}
