% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_one}
\alias{add_one}
\title{Add One to a Number}
\usage{
add_one(number)
}
\arguments{
\item{number}{Numeric. A number to which one will be added. See details.}
}
\value{
Numeric.
}
\description{
Add One to a Number
}
\details{
At Member Level 1, you'll get your supplied number plus 1. Except 20\% of the
time it'll be wrong!

At Member Level 2, the answer will always be correct!

At Member Level 3, the answer will print in a colour of your choice!

If you run out of MEGACOINS, simply \link{watch_ad}!
}
\examples{
\dontrun{add_one(1)}
}
